requirements.txt
asgiref==3.2.10
Django==3.0.8
django-appconf==1.0.4
django-filter==2.3.0
Django-Select2==5.8.10
django-cors-headers==3.4.0
django-import_export
https://github.com/darklow/django-suit/tarball/v2
djangorestframework==3.11.0
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
pillow
gunicorn==20.0.4
